==> 1_noaa_import.ipynb <==

# 1_noaa_import.ipynb

This file imports the raw NOAA temperature files (average, min, and max temperature for each month for each state), reshapes them, and creates the 5 main variables of interest:
- `pastyear`
- `year1`
- `year2`
- `year3`
- `year4`

# Libraries
import pandas as pd
import numpy as np

# Create month abbreviation to numeric number crosswalk
month = ['jan', 'feb', 'mar', 'apr', 'may', 'jun', 'jul', 'aug', 'sep', 'oct', 'nov', 'dec']
month_num = list(range(1,13))
months = pd.DataFrame({'month': month, 'month_num': month_num})

# Import fips code list
fips_codes = pd.read_csv(os.getcwd() + '/data/state_fips.txt', sep = '\t')
fips_codes['polyname'] = fips_codes['polyname'].str.lower()

# Import NOAA geography code list
noaa_codes_st = pd.read_csv(os.getcwd() + '/data/noaa_state_codes_st.csv')
noaa_codes_st['polyname'] = noaa_codes_st['noaa_geoname'].str.lower()
noaa_codes_st = noaa_codes_st[noaa_codes_st['noaa_geocode'] < 100]
noaa_codes_st = noaa_codes_st.merge(fips_codes, on = 'polyname', how = 'left')
noaa_codes_st = noaa_codes_st[['noaa_geocode', 'polyname', 'state_fips']]



# Read in NOAA temperature fixed width files
avgtemp_raw = pd.read_fwf(os.getcwd() + '/data/climdiv-tmpcst-v1.0.0-20200904',
                          names = ['noaa_geocode','division', 'element', 'year','jan', 'feb', 'mar', 'apr', 'may', 'jun', 'jul', 'aug', 'sep', 'oct', 'nov', 'dec'],
                          widths = [3, 1, 2, 4, 7, 7 , 7, 7, 7, 7, 7, 7, 7, 7, 7, 7 ]).drop(columns = ['division', 'element'])
avgtemp_raw = avgtemp_raw[avgtemp_raw['noaa_geocode'] < 100]

mintemp_raw = pd.read_fwf(os.getcwd() + '/data/climdiv-tminst-v1.0.0-20200904',
                          names = ['noaa_geocode','division', 'element', 'year','jan', 'feb', 'mar', 'apr', 'may', 'jun', 'jul', 'aug', 'sep', 'oct', 'nov', 'dec'],
                          widths = [3, 1, 2, 4, 7, 7 , 7, 7, 7, 7, 7, 7, 7, 7, 7, 7 ]).drop(columns = ['division', 'element'])
mintemp_raw = mintemp_raw[mintemp_raw['noaa_geocode'] < 100]

maxtemp_raw = pd.read_fwf(os.getcwd() + '/data/climdiv-tmaxst-v1.0.0-20200904',
                          names = ['noaa_geocode','division', 'element', 'year','jan', 'feb', 'mar', 'apr', 'may', 'jun', 'jul', 'aug', 'sep', 'oct', 'nov', 'dec'],
                          widths = [3, 1, 2, 4, 7, 7 , 7, 7, 7, 7, 7, 7, 7, 7, 7, 7 ]).drop(columns = ['division', 'element'])
maxtemp_raw = maxtemp_raw[maxtemp_raw['noaa_geocode'] < 100]

avgtemp_raw.head()

# Reshape each NOAA temperature dataset from wide to long
avgtemp_long_a = (avgtemp_raw.melt(id_vars=['noaa_geocode', 'year'], value_vars = month)
                             .rename(columns = {'variable':'month', 'value':'avg_temp'})
                             .merge(months, on = 'month', how = 'left'))

mintemp_long_a = (mintemp_raw.melt(id_vars=['noaa_geocode', 'year'], value_vars = month)
                             .rename(columns = {'variable':'month', 'value':'min_temp'})
                             .merge(months, on = 'month', how = 'left'))

maxtemp_long_a = (maxtemp_raw.melt(id_vars=['noaa_geocode', 'year'], value_vars = month)
                             .rename(columns = {'variable':'month', 'value':'max_temp'})
                             .merge(months, on = 'month', how = 'left'))

alltemp_long_a = avgtemp_long_a.merge(mintemp_long_a, on = ['noaa_geocode', 'month', 'month_num', 'year'], how = 'left')
alltemp_long_b = alltemp_long_a.merge(maxtemp_long_a, on = ['noaa_geocode', 'month', 'month_num', 'year'], how = 'left')
df_w_measures = alltemp_long_a.sort_values(['noaa_geocode', 'year', 'month_num'])
df_w_measures.head()

# Measure Creator: creates temperature variables for analysis
def measure_creator(measure, measure_name):

    df_w_measures = alltemp_long_b.copy().sort_values(['noaa_geocode', 'year', 'month_num'])

    df_w_measures["lead2"] = df_w_measures.groupby('noaa_geocode')[measure].shift(-2)
    df_w_measures["lead1"] = df_w_measures.groupby('noaa_geocode')[measure].shift(-1)
    df_w_measures["lag0"] = df_w_measures.groupby('noaa_geocode')[measure].shift(0)
    df_w_measures["lag1"] = df_w_measures.groupby('noaa_geocode')[measure].shift(1)
    df_w_measures["lag2"] = df_w_measures.groupby('noaa_geocode')[measure].shift(2)
    df_w_measures["lag3"] = df_w_measures.groupby('noaa_geocode')[measure].shift(3)
    df_w_measures["lag4"] = df_w_measures.groupby('noaa_geocode')[measure].shift(4)
    df_w_measures["lag5"] = df_w_measures.groupby('noaa_geocode')[measure].shift(5)
    df_w_measures["lag6"] = df_w_measures.groupby('noaa_geocode')[measure].shift(6)
    df_w_measures["lag7"] = df_w_measures.groupby('noaa_geocode')[measure].shift(7)
    df_w_measures["lag8"] = df_w_measures.groupby('noaa_geocode')[measure].shift(8)
    df_w_measures["lag9"] = df_w_measures.groupby('noaa_geocode')[measure].shift(9)
    df_w_measures["lag10"] = df_w_measures.groupby('noaa_geocode')[measure].shift(10)
    df_w_measures["lag11"] = df_w_measures.groupby('noaa_geocode')[measure].shift(11)
    df_w_measures["lag12"] = df_w_measures.groupby('noaa_geocode')[measure].shift(12)
    df_w_measures["lag24"] = df_w_measures.groupby('noaa_geocode')[measure].shift(24)
    df_w_measures["lag36"] = df_w_measures.groupby('noaa_geocode')[measure].shift(36)
    df_w_measures["lag48"] = df_w_measures.groupby('noaa_geocode')[measure].shift(48)
    df_w_measures["lag60"] = df_w_measures.groupby('noaa_geocode')[measure].shift(60)

    df_w_measures['first3months'] = df_w_measures[['lead2', 'lead1', 'lag0']].mean(axis=1)
    df_w_measures['trimester3'] = df_w_measures[['lag1', 'lag2', 'lag3']].mean(axis=1)
    df_w_measures['trimester2'] = df_w_measures[['lag4', 'lag5', 'lag6']].mean(axis=1)
    df_w_measures['trimester1'] = df_w_measures[['lag7', 'lag8', 'lag9']].mean(axis=1)
    df_w_measures['ninemonth'] = df_w_measures[['lag1', 'lag2', 'lag3', 'lag4', 'lag5', 'lag6', 'lag7', 'lag8', 'lag9']].mean(axis=1)
    df_w_measures['pastyear'] = df_w_measures[['lag1', 'lag2', 'lag3', 'lag4', 'lag5', 'lag6', 'lag7', 'lag8', 'lag9', 'lag10', 'lag11', 'lag12']].mean(axis=1)

    df_w_measures["year1"] = df_w_measures.groupby('noaa_geocode')['pastyear'].shift(-12)
    df_w_measures["year2"] = df_w_measures.groupby('noaa_geocode')['pastyear'].shift(-24)
    df_w_measures["year3"] = df_w_measures.groupby('noaa_geocode')['pastyear'].shift(-36)
    df_w_measures["year4"] = df_w_measures.groupby('noaa_geocode')['pastyear'].shift(-48)
    df_w_measures["year5"] = df_w_measures.groupby('noaa_geocode')['pastyear'].shift(-60)

    df_w_measures['pastyearx'] = df_w_measures[['lag0', 'lag1', 'lag2', 'lag3', 'lag4', 'lag5', 'lag6', 'lag7', 'lag8', 'lag9', 'lag10', 'lag11']].mean(axis=1)

    df_w_measures["year1x"] = df_w_measures.groupby('noaa_geocode')['pastyearx'].shift(-12)
    df_w_measures["year2x"] = df_w_measures.groupby('noaa_geocode')['pastyearx'].shift(-24)
    df_w_measures["year3x"] = df_w_measures.groupby('noaa_geocode')['pastyearx'].shift(-36)
    df_w_measures["year4x"] = df_w_measures.groupby('noaa_geocode')['pastyearx'].shift(-48)
    df_w_measures["year5x"] = df_w_measures.groupby('noaa_geocode')['pastyearx'].shift(-60)



    keep_same = {'noaa_geocode', 'year', 'month', 'month_num', 'avg_temp', 'min_temp', 'max_temp'}
    df_w_measures.columns = ['{}{}'.format(c, '' if c in keep_same else "_" + measure_name)
               for c in df_w_measures.columns]

    return df_w_measures

measures_max = measure_creator('max_temp', "MAX").copy()
measures_avg = measure_creator('avg_temp', "AVG").copy()
measures_min = measure_creator('min_temp', "MIN").copy()

measures_avg.head()

# Combine all measures_ datasets
alltemp_long_c = measures_avg.merge(measures_min, on = ['noaa_geocode', 'year', 'month', 'month_num', 'avg_temp', 'min_temp', 'max_temp'], how = 'left')
alltemp_long_c = alltemp_long_c.merge(measures_max, on = ['noaa_geocode', 'year', 'month', 'month_num', 'avg_temp', 'min_temp', 'max_temp'], how = 'left')

# Drop lag and lead variables to simplify the DataFrame
alltemp_long_d = alltemp_long_c[alltemp_long_c.columns.drop(list(alltemp_long_c.filter(regex='lag')))]
alltemp_long_d = alltemp_long_d[alltemp_long_d.columns.drop(list(alltemp_long_d.filter(regex='lead')))]


alltemp_long_d.head()

alltemp_long_e = (alltemp_long_d.merge(noaa_codes_st, on = ['noaa_geocode'], how="left")
                                .rename(columns = {'month_num':'bmonth',
                                                   'month':'bmonth_name',
                                                   'year':'byear',
                                                   'polyname':'bpl_name',
                                                   'state_fips': 'BPL_int'}))
alltemp_long_e

alltemp_long_e.year1_AVG

measures_avg
