==> 2_bunmd_import.ipynb <==
# 2_bunmd_import.ipynb

This file reads in the the large BUNMD file and creates subset files for women and men.

import pandas as pd

# Read in BUNMD file -- note: large file 7.6GB -- best time so far is 3m 53.2s

bunmd2_raw = pd.read_csv(os.getcwd() + '/data/bunmd_v2.csv')

bunmd2_raw_male = bunmd2_raw[bunmd2_raw['sex'] == 1]
bunmd2_raw_male.to_csv(os.getcwd() + "/data/bunmd2_raw_male.csv")

bunmd2_raw_female = bunmd2_raw[bunmd2_raw['sex'] == 2]
bunmd2_raw_female.to_csv(os.getcwd() + "/data/bunmd2_raw_female.csv")
