==> 3_bunmd_prep.ipynb <==
# 3_bunmd_prep.ipynb

This file wrangles the gender-specific BUNMD files, including dropping NA weight values and merging in the NOAA temperature data.

import pandas as pd
import numpy as np

%run -i '1_noaa_import.ipynb'

bunmd2_raw_female = pd.read_csv(os.getcwd() + '/data/bunmd2_raw_female.csv')
bunmd2_raw_male = pd.read_csv(os.getcwd() + '/data/bunmd2_raw_male.csv')

def bunmd_wrangle(dataset):
    dataset = dataset[dataset['weight'].notna()]
    dataset['bpl_name'] = dataset['bpl_string'].str.lower()
    dataset = dataset.merge(alltemp_long_e, on = ["byear", "bmonth", "bpl_name"], how="left")
    dataset = dataset[dataset['avg_temp'].notna()]
    return dataset

bunmd2_male_CLEAN = bunmd_wrangle(bunmd2_raw_male)

bunmd2_male_CLEAN.to_csv(os.getcwd() + "/data/bunmd2_male_CLEAN.csv")

bunmd2_female_CLEAN = bunmd_wrangle(bunmd2_raw_female)

bunmd2_female_CLEAN.to_csv(os.getcwd() + "/data/bunmd2_female_CLEAN.csv")
